+++ apps/angular-swa-auth-demo/src/app/app.module.ts
@@ -5,6 +5,7 @@ import { RouterModule } from '@angular/router';
 import { IdentityProviderInteractiveSelectorService, SwaAuthModule } from '@christianacca/angular-swa-auth';
 import { environment } from '../environments/environment';
 import { AboutComponent } from './about.component';
+import { AdminLandingComponent } from './admin-landing.component';
 import { AppComponent } from './app.component';
 import { externalModules } from './build-specific';
 import { declarations } from './core';
@@ -12,7 +13,7 @@ import { routes } from './router';
 import { AppStoreModule } from './store/store.module';
 
 @NgModule({
-  declarations: [AppComponent, AboutComponent, declarations],
+  declarations: [AppComponent, AboutComponent, AdminLandingComponent, declarations],
   imports: [
     BrowserModule,
     HttpClientModule,