+++ libs/demo-app/product-admin/src/lib/containers/index.ts
@@ -1,0 +1,6 @@
+import { ProductCategoriesComponent } from './product-categories.component';
+import { ProductStatsComponent } from './product-stats.component';
+
+export { ProductCategoriesComponent, ProductStatsComponent };
+
+export const CONTAINERS = [ProductCategoriesComponent, ProductStatsComponent];