+++ libs/demo-app/user-admin/src/lib/containers/index.ts
@@ -1,0 +1,5 @@
+import { UserManagementComponent } from './user-management.component';
+import { UserStatsComponent } from './user-stats.component';
+
+export { UserManagementComponent, UserStatsComponent };
+export const CONTAINERS = [UserManagementComponent, UserStatsComponent];