+++ libs/demo-app/user-admin/src/lib/containers/user-management.component.ts
@@ -1,0 +1,8 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  template: ` <p>user-management works!</p> `,
+  styles: [],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class UserManagementComponent {}